address Mikhail's feedback
