+++ Explorer/Assets/DCL/Infrastructure/CrdtEcsBridge/JsModulesImplementation/Communications/SceneCommunicationPipe.cs
@@ -9,6 +9,7 @@ using LiveKit.Proto;
 using System;
 using System.Collections.Generic;
 using System.Threading;
+using DCL.SceneBannedUsers;
 
 namespace CrdtEcsBridge.JsModulesImplementation.Communications
 {
@@ -36,6 +37,7 @@ namespace CrdtEcsBridge.JsModulesImplementation.Communications
             {
                 ReadOnlySpan<byte> decodedMessage = message.Payload.Data.Span;
                 ISceneCommunicationPipe.MsgType msgType = DecodeMessage(ref decodedMessage);
+                bool isTrustedSource = IsTrustedSource(message.FromWalletId);
 
                 if (decodedMessage.Length == 0)
                     return;
@@ -52,8 +54,26 @@ namespace CrdtEcsBridge.JsModulesImplementation.Communications
                     if (!sceneMessageHandlers.TryGetValue(key, out handler))
                         return;
 
-                handler(new ISceneCommunicationPipe.DecodedMessage(decodedMessage, message.FromWalletId));
+                ISceneCommunicationPipe.DecodedMessage dm = new (
+                    decodedMessage,
+                    message.FromWalletId,
+                    isTrustedSource
+                );
+
+                handler(dm);
             }
+        }
+
+        private bool IsTrustedSource(string walletId)
+        {
+            SceneAdminResult result = RoomMetadataCurrentScene.Instance.IsAdmin(walletId);
+
+            return result.Match(
+                    onSuccess: () => true, // Message is considered safe if it's from a scene admin
+                    onLocalSceneDevelopment: () => true, //sceneAdmins are not applicable in cases like LSD
+                    onNotAdmin: () => false,
+                    onNotLoadedYet: () => false // Consider the user as non-admin until we know for sure
+                    );
         }
 
         private static ISceneCommunicationPipe.MsgType DecodeMessage(ref ReadOnlySpan<byte> value)
